DRTII-620 Refactor to tidy
