+++ server/src/test/scala/services/PcpArrivalSpec.scala
@@ -3,8 +3,9 @@ package services
 import controllers.ArrivalGenerator.arrival
 import drt.shared.Terminals.{T1, T2, Terminal}
 import drt.shared.api.{Arrival, WalkTime}
-import drt.shared.coachTime.{CoachTransfer, CoachWalkTime, DefaultCoachWalkTime, LhrCoachWalkTime}
-import drt.shared.{LhrRedListDatesImpl, MilliDate, PortCode}
+import drt.shared.coachTime.{CoachTransfer, DefaultCoachWalkTime, LhrCoachWalkTime}
+import drt.shared.redlist.LhrRedListDatesImpl
+import drt.shared.{MilliDate, PortCode}
 import org.specs2.mutable.SpecificationLike
 
 class PcpArrivalSpec extends SpecificationLike {
