+++ shared/src/main/scala/drt/shared/FlightDisplayFilter.scala
@@ -1,7 +1,7 @@
 package drt.shared
 
-import drt.shared.CrunchApi.MillisSinceEpoch
 import drt.shared.Terminals._
+import drt.shared.redlist.LhrTerminalTypes
 
 sealed trait FlightDisplayFilter {
   def forTerminal(flights: Iterable[ApiFlightWithSplits], terminal: Terminal): Iterable[ApiFlightWithSplits]
@@ -10,29 +10,6 @@ sealed trait FlightDisplayFilter {
 case object DefaultFlightDisplayFilter extends FlightDisplayFilter {
   def forTerminal(flights: Iterable[ApiFlightWithSplits], terminal: Terminal): Iterable[ApiFlightWithSplits] =
     flights.filter(_.apiFlight.Terminal == terminal)
-}
-
-trait LhrRedListDates {
-  val t3RedListOpeningDate: MillisSinceEpoch
-  val t4RedListOpeningDate: MillisSinceEpoch
-}
-
-case object LhrRedListDatesImpl extends LhrRedListDates {
-  override val t3RedListOpeningDate = 1622502000000L // 2021-06-01 BST
-  override val t4RedListOpeningDate = 1624921200000L // 2021-06-29 BST
-}
-
-case class LhrTerminalTypes(lhrRedListDates: LhrRedListDates) {
-
-  def lhrRedListTerminalForDate(scheduled: MillisSinceEpoch): Option[Terminal] =
-    if (scheduled < lhrRedListDates.t3RedListOpeningDate) None
-    else if (scheduled < lhrRedListDates.t4RedListOpeningDate) Option(T3)
-    else Option(T4)
-
-  def lhrNonRedListTerminalsForDate(scheduled: MillisSinceEpoch): List[Terminal] =
-    if (scheduled < lhrRedListDates.t3RedListOpeningDate) List()
-    else if (scheduled < lhrRedListDates.t4RedListOpeningDate) List(T2, T5)
-    else List(T2, T3, T5)
 }
 
 case class LhrFlightDisplayFilter(isRedListOrigin: PortCode => Boolean, terminalTypes: LhrTerminalTypes) extends FlightDisplayFilter {
